# openapi.yaml
openapi: 3.0.3
info:
  title: 辰星认证中枢 API
  version: 1.0.0
  description: |
    辰星通行证用户认证、OAuth 2.0/OIDC 和统一用户管理 API。
    当前定义以仓库实际路由为准。运行期 Issuer 持久化在 PostgreSQL；未配置时仅健康检查、静态前端和初始化状态可用，其他协议路由由运行时门禁拒绝。

    请求超时（`REQUEST_TIMEOUT_SECONDS`）按路径分流错误信封（Issue #423）：
    `/oauth/*` 协议端点返回 RFC 6749 `503 temporarily_unavailable`（`OAuthErrorResponse`），
    不得改写为内部 API 信封；其余应用路由返回 `504` + `ErrorResponse`（`code=request_timeout`）。
    健康检查与静态 SPA fallback 不套用该请求超时。
servers:
  - url: http://127.0.0.1:3000
    description: 本地开发
...
paths:
...
  /oauth/authorize:
    get:
      tags: [OAuth/OIDC]
      summary: OAuth 授权码入口
      operationId: authorize
      security: [{ sessionCookie: [] }, { chenxingSession: [] }, {}]
      description: |
        未登录浏览器请求仅在 client、redirect_uri、scope、state 和 PKCE 完整校验后创建 pending。
        X-Chenxing-Session 仅在 OAUTH_SESSION_HEADER_ENABLED=true 的开发兼容配置中接受，不能绕过 Session 绑定、CSRF 或 consent。
        请求超时在中间件层返回 503 `temporarily_unavailable`（RFC 6749 JSON 信封；此时尚无可用的 redirect_uri 上下文做错误重定向），不会返回内部 API 的 504 `request_timeout`（Issue #423）。
      parameters:
        - { $ref: '#/components/parameters/ClientId' }
        - { name: redirect_uri, in: query, required: true, schema: { type: string, format: uri } }
        - { name: response_type, in: query, required: true, schema: { type: string, enum: [code] } }
        - { name: scope, in: query, required: true, description: "空格分隔的 scope；每个值必须同时属于服务端 OAUTH_CLIENT_ALLOWED_SCOPES allowlist（默认 openid、profile、email）和该 Client 已注册的 scope 集合", schema: { type: string, example: 'openid email profile' } }
        - { name: state, in: query, required: true, description: 最多 512 个字符, schema: { type: string, maxLength: 512 } }
        - { name: nonce, in: query, description: 最多 512 个字符, schema: { type: string, maxLength: 512 } }
        - { name: code_challenge, in: query, required: true, schema: { type: string } }
        - { name: code_challenge_method, in: query, required: true, schema: { type: string, enum: [S256] } }
      responses:
        '302': { description: 重定向到 Client 的 redirect_uri 或登录/授权确认页（跳转到 SPA 登录页或授权确认页时附带授权持有者 HttpOnly Cookie；HTTPS 名称为 __Host-chenxing_authz_holder，本地 HTTP 名称为 chenxing_authz_holder，供后续绑定端点校验；已授权直通到 redirect_uri 的分支不下发） }
        '400': { $ref: '#/components/responses/OAuthBadRequest' }
        '401': { $ref: '#/components/responses/OAuthUnauthorized' }
        '429': { $ref: '#/components/responses/OAuthTooManyRequests' }
        '503': { $ref: '#/components/responses/OAuthServiceUnavailable' }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
    adminBearer: { type: http, scheme: bearer, description: 系统 ADMIN_TOKEN 自动化通道，权限等价于 Owner 且无用户 ID，不需要浏览器 CSRF。ADMIN_TOKEN 为空时整个管理面关闭：Bearer 与 sessionCookie 两条通道都被拒绝，已初始化的管理接口统一返回 403 admin_disabled；唯一例外是不存在 Owner 时公开的首个 Owner 初始化接口 }
    basicClientAuth: { type: http, scheme: basic }
    sessionCookie: { type: apiKey, in: cookie, name: __Host-chenxing_session, description: HTTPS 浏览器部署使用的 host-only Session Cookie；仅 loopback HTTP 本地开发使用不带前缀的兼容名称 }
    chenxingSession: { type: apiKey, in: header, name: X-Chenxing-Session, description: 仅开发兼容配置 OAUTH_SESSION_HEADER_ENABLED=true 时可用；不能替代 pending、Session 绑定、CSRF 或 consent }
  parameters:
    ClientId: { name: client_id, in: query, required: true, schema: { type: string } }
...
  responses:
...
    OAuthBadRequest: { description: OAuth 请求错误，响应遵守 RFC 6749/7009, content: { application/json: { schema: { $ref: '#/components/schemas/OAuthErrorResponse' } } } }
    OAuthUnauthorized:
      description: OAuth 客户端或 Bearer 凭据无效
      headers:
        WWW-Authenticate: { description: RFC 6749/6750 challenge, schema: { type: string } }
      content: { application/json: { schema: { $ref: '#/components/schemas/OAuthErrorResponse' } } }
    OAuthTooManyRequests: { description: OAuth 限流或容量保护触发, content: { application/json: { schema: { $ref: '#/components/schemas/OAuthErrorResponse' } } } }
    OAuthServiceUnavailable:
      description: |
        OAuth 依赖暂时不可用，或 `/oauth/*` 协议端点触发请求超时（Issue #423）。
        响应始终为 RFC 6749 信封（`error=temporarily_unavailable`），不是内部 `{code, message}`。
      content: { application/json: { schema: { $ref: '#/components/schemas/OAuthErrorResponse' } } }
...
  schemas:
...
    OAuthErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string, description: RFC 6749/6750/7009 error code }
        error_description: { type: string, description: Human-readable non-sensitive description }
...